Float to decimal

+++ ExpenseTrackingApplication/ViewModels/EditTransactionViewModel.cs
@@ -6,7 +6,7 @@ public class EditTransactionViewModel
 {
     public int? Id { get; set; }
     public string? Recipient { get; set; }
-    public float Amount { get; set; }
+    public decimal Amount { get; set; }
     public DateTime Date { get; set; }
     public TransactionCategory Category { get; set; }
     public string? Description { get; set; }
